Projects/02 Posts Website/requirements.txt
Bootstrap-Flask==2.3.3
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.4.2
docopt==0.6.2
email-validator==2.1.0.post1
Flask==2.2.5
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.2.1
greenlet==3.0.1
idna==3.6
itsdangerous==2.0.1
Jinja2==3.1.2
loguru==0.7.2
MarkupSafe==2.1.1
Pillow==10.1.0
pip==23.3.1
pipreqs==0.4.13
requests==2.31.0
setuptools==68.2.2
SQLAlchemy==2.0.21
typing_extensions==4.7.1
urllib3==2.1.0
Werkzeug==2.2.3
wheel==0.41.2
WTForms==3.1.1
yarg==0.1.9
